requirements/requirements-testing.txt
# Pytest for running the tests.
pytest==5.1.2
pytest-django==3.5.1
pytest-cov==2.7.1

